requirements.txt
djangorestframework==3.13.1
django-money==2.1.1
django-phonenumber-field==6.1.0
Django==4.0.3
drf-yasg==1.20.0
phonenumbers==8.12.44
pytz==2021.3
python-dateutil==2.8.2
sqlparse==0.4.2
pytest==7.0.1
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-sugar==0.9.4
pre-commit==2.17.0  # https://github.com/pre-commit/pre-commit
